+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Fix attribute access to stub-only names (#525)
 - Remove a number of unnecessary special-cased signatures
   (#499)
 - Add support for use of the `Unpack` operator to