+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Fix resolution of stringified annotations in `__init__` methods (#530)
 - Type check `yield`, `yield from`, and `return` nodes in generators (#529)
 - Type check calls to comparison operators (#527)
 - Retrieve attributes from stubs even when a runtime